requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
python-dotenv
pymysql
python-dotenv
